packages/control-rod/control_rod-2024.8.2.0.tar.gz/control_rod-2024.8.2.0/abobjs/control.py
#!/usr/bin/env python3

import logging
import urllib.parse
import datetime
import json

import requests
import pyjq

import abobjs

class AB_Control(abobjs.AB_Base):

    _endpoint = "/api/v1/controls"
    _single = ".controls[0]"
    _name = "control"

    def __init__(self, id=None, api_info=None, **kwargs):

        # Dunder Handling
        endpoint = kwargs.get("endpoint", self._endpoint)
        single = kwargs.get("single", self._single)
        name = kwargs.get("name", self._name)

        abobjs.AB_Base.__init__(self, id=id, api_info=api_info,
                                endpoint=endpoint, single=single, name=name,
                                **kwargs)
